bbftpd: Added -A option to tell it to ask the client for its IP address when started via ssh. The way bbftpd normally finds the client's IP address is that it opens a port a requests that the client connect to that port.  Once connected, the client's IP address becomes known to the server.  However, this fails when the server is located behind a firewall that prevents the client from connecting.  The -A option tells the server to ask the client for its IP address.
In adding this feature, I totally restructured bbftpd.c to make it
easier to understand and maintain.  I also added additional functions
that wrap the low-level read/write functions and used the new
functions in the updated code.  The addition of these functions adds
much more clarity to the code for better maintenance.